requirements.txt
# Core
pvporcupine>=3.0.0
pyaudio>=0.2.14
numpy>=1.26.0
httpx>=0.27.0
pyyaml>=6.0.1
soundfile>=0.12.1

# STT
openai>=1.30.0

# VAD
torch>=2.1.0
torchaudio>=2.1.0
silero-vad>=5.1

# TTS
piper-tts>=1.2.0

# Utilities
python-dotenv>=1.0.0
